Break out SessionDatum into subclasses
