+++ tests/test/org/commcare/backend/session/test/SessionNavigatorTests.java
@@ -1,5 +1,6 @@
 package org.commcare.backend.session.test;
 
+import org.commcare.suite.model.EntityDatum;
 import org.commcare.suite.model.SessionDatum;
 import org.commcare.test.utilities.MockSessionNavigationResponder;
 import org.commcare.test.utilities.MockApp;
@@ -90,7 +91,7 @@ public class SessionNavigatorTests {
 
         // Confirm that the next datum has auto-select enabled
         SessionDatum nextNeededDatum = session.getNeededDatum();
-        Assert.assertTrue(nextNeededDatum.isAutoSelectEnabled());
+        Assert.assertTrue(((EntityDatum)nextNeededDatum).isAutoSelectEnabled());
 
         // Since there are 2 cases in the case list (user_restore.xml contains 2 cases of type
         // 'pregnancy', which is the case type that m0-f2's nodeset filters for), the
@@ -112,7 +113,7 @@ public class SessionNavigatorTests {
 
         // Confirm that the next datum has auto-select enabled and has a confirm detail defined
         SessionDatum nextNeededDatum = session.getNeededDatum();
-        Assert.assertTrue(nextNeededDatum.isAutoSelectEnabled());
+        Assert.assertTrue(((EntityDatum)nextNeededDatum).isAutoSelectEnabled());
         Assert.assertNotNull(nextNeededDatum.getLongDetail());
 
         // Since there is one case in the case list (user_restore.xml contains one case of type
@@ -135,7 +136,7 @@ public class SessionNavigatorTests {
         // Confirm that the next datum has auto-select enabled, but does NOT have a confirm detail
         // defined
         SessionDatum nextNeededDatum = session.getNeededDatum();
-        Assert.assertTrue(nextNeededDatum.isAutoSelectEnabled());
+        Assert.assertTrue(((EntityDatum)nextNeededDatum).isAutoSelectEnabled());
         Assert.assertNull(nextNeededDatum.getLongDetail());
 
         // Since there is one case in the case list and auto-select is enabled, but there is no
